+++ src/Plugin/HipchatNotify.php
@@ -45,7 +45,7 @@ class HipchatNotify extends Plugin
         }
 
         if (\array_key_exists('auth_token', $options)) {
-            $this->authToken = $this->builder->interpolate($options['auth_token']);
+            $this->authToken = $this->builder->interpolate($options['auth_token'], true);
         }
 
         $this->room = $options['room'];